+++ openedx_filters/authentication/__init__.py
@@ -1,0 +1,6 @@
+"""
+Filters related to the authentication subdomain.
+
+The authentication subdomain covers how users sign in, register, and are routed through
+login and registration flows.
+"""